Reconnect on DB failure

+++ process_messages.py
@@ -2,11 +2,27 @@ import json
 import time
 import boto3
 from config import *
+import mysql.connector
 
 sqs = boto3.client('sqs')
+db = None
+cursor = None
+
+
+def attempt_db_connection():
+    global db, cursor
+    try:
+        db = mysql.connector.connect(**SSL_CONFIG)
+        print("Connected to DB!")
+        return True
+    except Exception as e:
+        print(e)
+        return False
 
 
 def process_queue():
+    global db, cursor
+
     # Receive message from SQS queue
     response = sqs.receive_message(
         QueueUrl=queue_url,
@@ -21,22 +37,18 @@ def process_queue():
         WaitTimeSeconds=0
     )
 
-    # print(response)
     if "Messages" not in response.keys():
         print("No messages to process.")
         return False
 
-    messages = response['Messages']
-    print(f"message count: {len(messages)}")
     message = response['Messages'][0]
     body = json.loads(message['Body'])
     receipt_handle = message['ReceiptHandle']
     message_id = message['MessageId']
 
     print(f"Received message: {message_id}")
-    # print(f"{body}")
 
-    if store_log(body):
+    if db and store_log(body):
         sqs.delete_message(
             QueueUrl=queue_url,
             ReceiptHandle=receipt_handle
@@ -47,36 +59,33 @@ def process_queue():
         return False
 
 
-# body = {
-#     'action': 'insert',
-#     'table': 'canary_metrics',
-#     'values': {
-#         'path': '/logs',
-#         'ms_elapsed': 82.58699999999999,
-#         'timestamp': '2020-12-24 14:34:19.940642',
-#         'success': True
-#     }
-# }
-
 def store_log(body):
+    global db, cursor
     try:
+        db.ping(True)  # Assert connection valid, otherwise attempt to reconnect (True)
+
         path = body["values"]["path"]
         success = body["values"]["success"]
         timestamp = body["values"]["timestamp"]
         elapsed_ms = body["values"]["ms_elapsed"]
 
+        cursor = db.cursor()  # Cursor init is very cheap, so we do not reuse it
         sql = f"INSERT INTO graphing_data.canary_metrics(path, ms_elapsed, timestamp, error, success) " \
               f"VALUES('{path}', '{elapsed_ms}', '{timestamp}', '{int(not success)}', '{int(success)}')"
         cursor.execute(sql)
         db.commit()
+        cursor.close()
         print(f"{sql}")
         return True
     except Exception as e:
-        print(f"Could not store metrics: {e}")
+        print(e)
         return False
 
 
 while True:
+    if not db:
+        attempt_db_connection()
+
     print(f"Checking queue...")
     if not process_queue():
         time.sleep(10)
